+++ Runtime/Scripts/Components/UIActivatorAnimations/LitMotions/UIMoveFromActiveAnimation.cs
@@ -7,26 +7,30 @@ namespace ParkMinPackages.UGUI.Components.UIActivatorAnimations.LitMotions
 {
 	[MovedFrom(true, "ParkMinPackages.UGUI.Components.UIActivatorAnimations", "ParkMinPackages.UGUI", "UILMMoveFromShowAnimation")]
 	[DisallowMultipleComponent]
-	public class UIMoveActiveAnimation : UILitMotionActiveAnimation
+	public class UIMoveFromActiveAnimation : UILitMotionActiveAnimation
 	{
 		public override MotionHandle CreateMotion() {
-			return LMotion.Create(HiddenPosition, ShownPosition, Duration)
+			return LMotion.Create(GetOffsetPosition(), _capturedPosition, Duration)
 			              .WithEase(Ease)
 			              .WithCancelOnError()
 			              .Bind(x => Target.localPosition = x)
 			              .AddTo(Target.gameObject);
 		}
-		public override void ApplyStart() {
-			Target.localPosition = HiddenPosition;
-		}
-		public override void ApplyEnd() {
-			Target.localPosition = ShownPosition;
-		}
-
 		public float Duration = 0.1f;
 		public Ease Ease = Ease.Linear;
-		public Vector3 ShownPosition;
-		public Vector3 HiddenPosition;
+		public Vector3 Offset;
+
+		protected override void CaptureValues() {
+			_capturedPosition = Target.localPosition;
+		}
+		protected override void RestoreCapturedValues() {
+			Target.localPosition = _capturedPosition;
+		}
+		Vector3 _capturedPosition;
+
+		Vector3 GetOffsetPosition() {
+			return _capturedPosition + Offset;
+		}
 	}
 }
 #endif