+++ Runtime/Scripts/Components/UIActivatorAnimations/LitMotions/UIScaleDeactivateAnimation.cs
@@ -2,6 +2,7 @@
 using LitMotion;
 using UnityEngine;
 using UnityEngine.Scripting.APIUpdating;
+using UnityEngine.Serialization;
 
 namespace ParkMinPackages.UGUI.Components.UIActivatorAnimations.LitMotions
 {
@@ -10,23 +11,24 @@ namespace ParkMinPackages.UGUI.Components.UIActivatorAnimations.LitMotions
 	public class UIScaleDeactivateAnimation : UILitMotionDeactivateAnimation
 	{
 		public override MotionHandle CreateMotion() {
-			return LMotion.Create(StartScale, EndScale, Duration)
+			return LMotion.Create(_capturedScale, Scale, Duration)
 			              .WithEase(Ease)
 			              .WithCancelOnError()
 			              .Bind(x => Target.localScale = x)
 			              .AddTo(Target.gameObject);
 		}
-		public override void ApplyStart() {
-			Target.localScale = StartScale;
-		}
-		public override void ApplyEnd() {
-			Target.localScale = EndScale;
-		}
-
 		public float Duration = 0.2f;
 		public Ease Ease = Ease.Linear;
-		public Vector3 StartScale = Vector3.one;
-		public Vector3 EndScale = Vector3.zero;
+		[FormerlySerializedAs("EndScale")]
+		public Vector3 Scale = Vector3.zero;
+
+		protected override void CaptureValues() {
+			_capturedScale = Target.localScale;
+		}
+		protected override void RestoreCapturedValues() {
+			Target.localScale = _capturedScale;
+		}
+		Vector3 _capturedScale;
 	}
 }
 #endif