+++ Runtime/Scripts/Enums/AnimationCancelBehaviour.cs
@@ -2,7 +2,6 @@ namespace ParkMinPackages.UGUI.Enums
 {
 	public enum AnimationCancelBehaviour
 	{
-		Stop,
 		Complete,
 		ResetToStart
 	}