+++ Runtime/Scripts/Enums/UIActivationState.cs
@@ -3,7 +3,6 @@ namespace ParkMinPackages.UGUI.Enums
 	public enum UIActivationState
 	{
 		Active,
-		Inactive,
-		Canceled
+		Inactive
 	}
 }